Stop calling group_key twice for each key
group_key is the most expensive part of the entire process.
We can use the Hash constructor that allows you to specify
a block to call when a key is missing, to initialize the value.

Brings average time of a sample size of 2000 down from 8.5s to 5.1s

time ./redis-audit.rb -h localhost -s 2000